feat: hoist useAgentChat into agents/chat/react with product wrappers
Move the canonical `useAgentChat` implementation, WebSocket chat transport,
and chat wire types out of `@cloudflare/ai-chat` and into a new shared
`agents/chat/react` entry. `@cloudflare/ai-chat/react` and the new
`@cloudflare/think/react` become thin wrappers over this shared core, so
Think apps no longer need to depend on `@cloudflare/ai-chat` just to get the
React hook.

Why
---
Using `useAgentChat` from `@cloudflare/ai-chat` against a Think server was
awkward and behaviorally divergent. The biggest footgun was `setMessages`:
`AIChatAgent` persists a client-pushed flat transcript, while Think is
server-authoritative and silently ignores it. Apps had to pull in an
unrelated package and still got surprising behavior.

What changed
------------
- agents: add `agents/chat/react` exposing `useAgentChat`,
  `WebSocketChatTransport`, and shared chat wire types
  (`MessageType`, `OutgoingMessage`, `IncomingMessage`). Add a new
  `syncMessagesToServer` option (default `true`) so server-authoritative
  hosts can keep `setMessages` local-only. `@ai-sdk/react` is added as an
  optional peer dependency (only the chat/react subpath needs it).
- ai-chat: `react.tsx`, `types.ts`, and `ws-chat-transport.ts` are now
  thin re-export shims over `agents/chat/react`. Public API and behavior
  are unchanged for existing consumers.
- think: add `@cloudflare/think/react`, a Think-tuned wrapper that hardcodes
  `syncMessagesToServer: false` (and omits it from its option type), so
  `setMessages` stays a local view update. A dev-only warn-once covers plain
  JS callers that pass the option anyway.
- think (server): warn once (dev only) when Think receives a client-pushed
  `CF_AGENT_CHAT_MESSAGES` frame, naming `@cloudflare/think/react` and
  `clearHistory()` as the fix. This is behavior-triggered, so it only fires
  when an app actually hits the divergence (e.g. the ai-chat hook against a
  Think server).
- docs/examples/starters: migrate all Think-backed clients to import the
  hook from `@cloudflare/think/react` and drop the now-unnecessary
  `@cloudflare/ai-chat` dependency. AIChatAgent-backed examples are
  unchanged. Document that `setMessages` is display-only on Think.

Changesets: agents (minor), @cloudflare/think (minor),
@cloudflare/ai-chat (patch).
