+++ .changeset/think-react-wrapper.md
@@ -1,0 +1,5 @@
+---
+"@cloudflare/think": minor
+---
+
+Add `@cloudflare/think/react`, a Think-tuned `useAgentChat` wrapper that keeps `setMessages` local-only by default while reusing the shared chat React implementation.
