+++ docs/think/client-tools.md
@@ -18,7 +18,7 @@ On the client, pass `clientTools` to `useAgentChat`:
 
 ```tsx
 import { useAgent } from "agents/react";
-import { useAgentChat } from "@cloudflare/ai-chat/react";
+import { useAgentChat } from "@cloudflare/think/react";
 
 function Chat() {
   const agent = useAgent({ agent: "MyAgent" });
