+++ examples/assistant/package.json
@@ -11,7 +11,6 @@
     "types": "think types"
   },
   "dependencies": {
-    "@cloudflare/ai-chat": "*",
     "@cloudflare/codemode": "*",
     "@cloudflare/kumo": "^2.5.2",
     "@cloudflare/shell": "*",
