+++ examples/chat-sdk-messenger/package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@chat-adapter/telegram": "^4.30.0",
-    "@cloudflare/ai-chat": "*",
     "@cloudflare/kumo": "^2.5.2",
     "@cloudflare/think": "*",
     "@phosphor-icons/react": "^2.1.10",
