+++ experimental/session-skills/package.json
@@ -10,7 +10,6 @@
     "types": "wrangler types env.d.ts --include-runtime false"
   },
   "dependencies": {
-    "@cloudflare/ai-chat": "*",
     "@cloudflare/kumo": "^2.5.2",
     "@cloudflare/think": "*",
     "@phosphor-icons/react": "^2.1.10",
