+++ packages/agents/src/chat/ws-chat-transport.ts
@@ -1,0 +1,787 @@
+/**
+ * WebSocket-based ChatTransport for useAgentChat.
+ *
+ * Replaces the aiFetch + DefaultChatTransport indirection with a direct
+ * WebSocket implementation that speaks the CF_AGENT protocol natively.
+ *
+ * Data flow (old): WS → aiFetch fake Response → DefaultChatTransport → useChat
+ * Data flow (new): WS → WebSocketChatTransport → useChat
+ */
+
+import type { ChatTransport, UIMessage, UIMessageChunk } from "ai";
+import { nanoid } from "nanoid";
+import { MessageType, type OutgoingMessage } from "./wire-types";
+
+/**
+ * Agent-like interface for sending/receiving WebSocket messages.
+ * Matches the shape returned by useAgent from agents/react.
+ */
+export interface AgentConnection {
+  send: (data: string) => void;
+  addEventListener: (
+    type: string,
+    listener: (event: MessageEvent) => void,
+    options?: { signal?: AbortSignal }
+  ) => void;
+  removeEventListener: (
+    type: string,
+    listener: (event: MessageEvent) => void
+  ) => void;
+}
+
+export type WebSocketChatTransportOptions<
+  ChatMessage extends UIMessage = UIMessage
+> = {
+  /** The agent connection from useAgent */
+  agent: AgentConnection;
+  /**
+   * Callback to prepare the request body before sending.
+   * Can add custom headers, body fields, or credentials.
+   */
+  prepareBody?: (options: {
+    messages: ChatMessage[];
+    trigger: "submit-message" | "regenerate-message";
+    messageId?: string;
+  }) => Promise<Record<string, unknown>> | Record<string, unknown>;
+  /**
+   * Optional set to track active request IDs.
+   * IDs are added when a request starts and removed when it completes.
+   * Used by the onAgentMessage handler to skip messages already handled by the transport.
+   */
+  activeRequestIds?: Set<string>;
+  /**
+   * Whether generic client-side abort/cancel lifecycle should cancel the
+   * server turn. Explicit cancellation via cancelActiveServerTurn() always
+   * sends CF_AGENT_CHAT_REQUEST_CANCEL.
+   * @default false
+   */
+  cancelOnClientAbort?: boolean;
+};
+
+/**
+ * ChatTransport that sends messages over WebSocket and returns a
+ * ReadableStream<UIMessageChunk> that the AI SDK's useChat consumes directly.
+ * No fake fetch, no Response reconstruction, no double SSE parsing.
+ */
+export class WebSocketChatTransport<
+  ChatMessage extends UIMessage = UIMessage
+> implements ChatTransport<ChatMessage> {
+  agent: AgentConnection;
+  private prepareBody?: WebSocketChatTransportOptions<ChatMessage>["prepareBody"];
+  private activeRequestIds?: Set<string>;
+  private cancelOnClientAbort: boolean;
+
+  // Pending resume resolver — set by reconnectToStream, called by
+  // handleStreamResuming when onAgentMessage sees CF_AGENT_STREAM_RESUMING.
+  private _resumeResolver: ((data: { id: string }) => void) | null = null;
+  // Pending "no stream" resolver — called by handleStreamResumeNone
+  // when onAgentMessage sees CF_AGENT_STREAM_RESUME_NONE.
+  private _resumeNoneResolver: (() => void) | null = null;
+  // Set when a client-side tool result/approval is expected to trigger
+  // a new continuation stream. In this mode reconnectToStream() returns
+  // a deferred ReadableStream immediately so AI SDK status can transition
+  // to "submitted" before the server starts streaming.
+  private _expectToolContinuation = false;
+  private _abortToolContinuation: (() => boolean) | null = null;
+  private _activeServerTurnId: string | null = null;
+  private _cancelAttachedStream: (() => boolean) | null = null;
+
+  constructor(options: WebSocketChatTransportOptions<ChatMessage>) {
+    this.agent = options.agent;
+    this.prepareBody = options.prepareBody;
+    this.activeRequestIds = options.activeRequestIds;
+    this.cancelOnClientAbort = options.cancelOnClientAbort ?? false;
+  }
+
+  setCancelOnClientAbort(cancelOnClientAbort: boolean) {
+    this.cancelOnClientAbort = cancelOnClientAbort;
+  }
+
+  /**
+   * Explicitly cancel the active server turn, if any.
+   * This is separate from generic client-side abort/cancel lifecycle so
+   * clients can detach locally without stopping server work.
+   */
+  cancelActiveServerTurn(): boolean {
+    const requestId = this._activeServerTurnId;
+    let cancelledRequest = false;
+
+    if (requestId) {
+      this.sendCancelFrame(requestId);
+      this._cancelAttachedStream?.();
+      this.clearActiveServerTurn(requestId);
+      cancelledRequest = true;
+    }
+
+    const cancelledToolContinuation = this.abortActiveToolContinuation();
+    return cancelledRequest || cancelledToolContinuation;
+  }
+
+  private sendCancelFrame(requestId: string) {
+    try {
+      this.agent.send(
+        JSON.stringify({
+          id: requestId,
+          type: MessageType.CF_AGENT_CHAT_REQUEST_CANCEL
+        })
+      );
+    } catch {
+      // Ignore failures (e.g. agent already disconnected)
+    }
+  }
+
+  private setActiveServerTurn(
+    requestId: string,
+    cancelAttachedStream: (() => boolean) | null
+  ) {
+    this._activeServerTurnId = requestId;
+    this._cancelAttachedStream = cancelAttachedStream;
+  }
+
+  private clearActiveServerTurn(requestId: string) {
+    if (this._activeServerTurnId === requestId) {
+      this._activeServerTurnId = null;
+      this._cancelAttachedStream = null;
+    }
+  }
+
+  /**
+   * Mark that the next reconnectToStream() call should attach to a
+   * server-initiated tool continuation rather than a page-load resume.
+   */
+  expectToolContinuation() {
+    this._expectToolContinuation = true;
+  }
+
+  /**
+   * Abort the active client-side tool continuation stream, if one is attached
+   * to a server request id.
+   */
+  abortActiveToolContinuation(): boolean {
+    return this._abortToolContinuation?.() ?? false;
+  }
+
+  /**
+   * True when the transport is waiting for a resume handshake.
+   */
+  isAwaitingResume(): boolean {
+    return this._resumeResolver !== null || this._resumeNoneResolver !== null;
+  }
+
+  /**
+   * Called by onAgentMessage when it receives CF_AGENT_STREAM_RESUMING.
+   * If reconnectToStream is waiting, this handles the resume handshake
+   * (ACK + stream creation) and returns true. Otherwise returns false
+   * so the caller can use its own fallback path.
+   */
+  handleStreamResuming(data: { id: string }): boolean {
+    if (!this._resumeResolver) return false;
+    this._resumeResolver(data);
+    return true;
+  }
+
+  /**
+   * Called by onAgentMessage when it receives CF_AGENT_STREAM_RESUME_NONE.
+   * If reconnectToStream is waiting, resolves the promise with null
+   * immediately (no 5-second timeout). Returns true if handled.
+   */
+  handleStreamResumeNone(): boolean {
+    if (!this._resumeNoneResolver) return false;
+    this._resumeNoneResolver();
+    return true;
+  }
+
+  /**
+   * Called by the hook's shared message handler when a server turn finishes
+   * outside the currently attached transport stream, such as after local-only
+   * client cleanup.
+   */
+  handleServerTurnCompleted(requestId: string) {
+    this.clearActiveServerTurn(requestId);
+  }
+
+  /**
+   * Register a server turn that is being rendered outside a transport-owned
+   * stream, such as the hook's fallback cross-tab/resume observer path.
+   */
+  observeServerTurn(requestId: string) {
+    this.setActiveServerTurn(requestId, null);
+  }
+
+  async sendMessages(options: {
+    chatId: string;
+    messages: ChatMessage[];
+    abortSignal: AbortSignal | undefined;
+    trigger: "submit-message" | "regenerate-message";
+    messageId?: string;
+    body?: object;
+    headers?: Record<string, string> | Headers;
+    metadata?: unknown;
+  }): Promise<ReadableStream<UIMessageChunk>> {
+    const requestId = nanoid(8);
+    const abortController = new AbortController();
+    let completed = false;
+    let requestSent = false;
+
+    // Build the request body
+    let extraBody: Record<string, unknown> = {};
+    if (this.prepareBody) {
+      extraBody = await this.prepareBody({
+        messages: options.messages,
+        trigger: options.trigger,
+        messageId: options.messageId
+      });
+    }
+    if (options.body) {
+      extraBody = {
+        ...extraBody,
+        ...(options.body as Record<string, unknown>)
+      };
+    }
+
+    const bodyPayload = JSON.stringify({
+      messages: options.messages,
+      trigger: options.trigger,
+      ...extraBody
+    });
+
+    // Track this request so the onAgentMessage handler skips it
+    this.activeRequestIds?.add(requestId);
+
+    // Create a ReadableStream<UIMessageChunk> that emits parsed chunks
+    // as they arrive over the WebSocket
+    const agent = this.agent;
+    const activeIds = this.activeRequestIds;
+
+    // Single cleanup helper — every terminal path (done, error, abort)
+    // goes through here exactly once.
+    // keepId: when true, do NOT remove requestId from activeIds. Used by
+    // explicit cancellation so onAgentMessage skips in-flight chunks
+    // and the server's final done:true signal until cleanup happens there.
+    const finish = (
+      action: () => void,
+      keepId = false,
+      clearServerTurn = true
+    ) => {
+      if (completed) return;
+      completed = true;
+      if (clearServerTurn) {
+        this.clearActiveServerTurn(requestId);
+      }
+      try {
+        action();
+      } catch {
+        // Stream may already be closed
+      }
+      if (!keepId) {
+        activeIds?.delete(requestId);
+      }
+      abortController.abort();
+    };
+
+    const abortError = new Error("Aborted");
+    abortError.name = "AbortError";
+
+    const cancelActiveRequest = () => {
+      if (completed) return false;
+      finish(() => streamController.error(abortError), true);
+      return true;
+    };
+    this.setActiveServerTurn(requestId, cancelActiveRequest);
+
+    // Abort handler: terminate the local stream. By default, generic AI SDK
+    // abort/cancel lifecycle is local-only so durable server turns can continue
+    // and be resumed. Use cancelActiveServerTurn() for explicit user/app
+    // cancellation, or cancelOnClientAbort for request-lifetime semantics.
+    const onAbort = () => {
+      if (completed) return;
+      if (this.cancelOnClientAbort) {
+        if (requestSent) {
+          this.sendCancelFrame(requestId);
+        }
+        finish(() => streamController.error(abortError), requestSent);
+      } else {
+        finish(() => streamController.error(abortError), false, !requestSent);
+      }
+    };
+
+    // streamController is assigned synchronously by start(), so it is
+    // always available by the time onAbort or onMessage can fire.
+    let streamController!: ReadableStreamDefaultController<UIMessageChunk>;
+
+    const stream = new ReadableStream<UIMessageChunk>({
+      start(controller) {
+        streamController = controller;
+
+        const onMessage = (event: MessageEvent) => {
+          try {
+            const data = JSON.parse(
+              event.data as string
+            ) as OutgoingMessage<ChatMessage>;
+
+            if (data.type !== MessageType.CF_AGENT_USE_CHAT_RESPONSE) return;
+            if (data.id !== requestId) return;
+
+            if (data.error) {
+              finish(() =>
+                controller.error(new Error(data.body || "Stream error"))
+              );
+              return;
+            }
+
+            // Parse the body as UIMessageChunk and enqueue
+            if (data.body?.trim()) {
+              try {
+                const chunk = JSON.parse(data.body) as UIMessageChunk;
+                controller.enqueue(chunk);
+              } catch {
+                // Skip malformed chunk bodies
+              }
+            }
+
+            if (data.done) {
+              finish(() => controller.close());
+            }
+          } catch {
+            // Ignore non-JSON messages
+          }
+        };
+
+        const onClose = () => {
+          finish(() => controller.close(), false, false);
+        };
+
+        agent.addEventListener("message", onMessage, {
+          signal: abortController.signal
+        });
+        agent.addEventListener("close", onClose, {
+          signal: abortController.signal
+        });
+      },
+      cancel() {
+        onAbort();
+      }
+    });
+
+    // Handle abort from the caller
+    if (options.abortSignal) {
+      options.abortSignal.addEventListener("abort", onAbort, { once: true });
+      if (options.abortSignal.aborted) onAbort();
+    }
+
+    if (completed) {
+      return stream;
+    }
+
+    // Send the request over WebSocket
+    requestSent = true;
+    agent.send(
+      JSON.stringify({
+        id: requestId,
+        init: {
+          method: "POST",
+          body: bodyPayload
+        },
+        type: MessageType.CF_AGENT_USE_CHAT_REQUEST
+      })
+    );
+
+    return stream;
+  }
+
+  async reconnectToStream(_options: {
+    chatId: string;
+  }): Promise<ReadableStream<UIMessageChunk> | null> {
+    if (this._expectToolContinuation) {
+      this._expectToolContinuation = false;
+      return this._createToolContinuationStream();
+    }
+
+    // Detect whether the server has an active stream for this chat.
+    // Instead of registering our own addEventListener listener (which
+    // races with onAgentMessage), we set _resumeResolver so that
+    // onAgentMessage can call handleStreamResuming() synchronously
+    // when it sees CF_AGENT_STREAM_RESUMING — eliminating the race.
+    const activeIds = this.activeRequestIds;
+
+    return new Promise<ReadableStream<UIMessageChunk> | null>((resolve) => {
+      let resolved = false;
+      let timeout: ReturnType<typeof setTimeout> | undefined;
+
+      const done = (value: ReadableStream<UIMessageChunk> | null) => {
+        if (resolved) return;
+        resolved = true;
+        this._resumeResolver = null;
+        this._resumeNoneResolver = null;
+        if (timeout) clearTimeout(timeout);
+        resolve(value);
+      };
+
+      // Set the "no stream" resolver that handleStreamResumeNone() will call.
+      // When onAgentMessage sees CF_AGENT_STREAM_RESUME_NONE, it calls
+      // handleStreamResumeNone() which resolves immediately with null.
+      this._resumeNoneResolver = () => done(null);
+
+      // Set the resolver that handleStreamResuming() will call.
+      // When onAgentMessage sees CF_AGENT_STREAM_RESUMING, it calls
+      // handleStreamResuming() which invokes this callback.
+      this._resumeResolver = (data: { id: string }) => {
+        const requestId = data.id;
+
+        // Track this request so onAgentMessage skips subsequent chunks
+        activeIds?.add(requestId);
+
+        const stream = this._createResumeStream(requestId);
+
+        // Send ACK to server via the latest agent (the socket may
+        // have been replaced since reconnectToStream was called).
+        this.agent.send(
+          JSON.stringify({
+            type: MessageType.CF_AGENT_STREAM_RESUME_ACK,
+            id: requestId
+          })
+        );
+
+        // Return a ReadableStream fed by the replayed + live chunks
+        done(stream);
+      };
+
+      // Send the resume request. PartySocket queues sends when
+      // the socket isn't open yet and flushes on connect, so
+      // this works regardless of current readyState.
+      try {
+        this.agent.send(
+          JSON.stringify({
+            type: MessageType.CF_AGENT_STREAM_RESUME_REQUEST
+          })
+        );
+      } catch {
+        // WebSocket may already be closed
+      }
+
+      // Safety-net timeout: if the WebSocket never connects or the
+      // server is unreachable, resolve null. Under normal operation
+      // the server responds with STREAM_RESUMING or STREAM_RESUME_NONE
+      // well before this fires.
+      timeout = setTimeout(() => done(null), 5000);
+    });
+  }
+
+  /**
+   * Creates a deferred ReadableStream for client-side tool continuations.
+   * The stream is returned immediately so AI SDK status becomes "submitted"
+   * right after addToolOutput()/addToolApprovalResponse(), then it waits for
+   * the server to announce the continuation via STREAM_RESUMING.
+   */
+  private _createToolContinuationStream(): ReadableStream<UIMessageChunk> {
+    const agent = this.agent;
+    const activeIds = this.activeRequestIds;
+    const streamController = new AbortController();
+    const abortError = new Error("Aborted");
+    abortError.name = "AbortError";
+    let completed = false;
+    let requestId: string | null = null;
+    let readerController!: ReadableStreamDefaultController<UIMessageChunk>;
+    let onResumeRef: ((data: { id: string }) => void) | null = null;
+    let onResumeNoneRef: (() => void) | null = null;
+
+    const clearHandshakeResolvers = (
+      resumeResolver?: ((data: { id: string }) => void) | null,
+      resumeNoneResolver?: (() => void) | null
+    ) => {
+      if (resumeResolver === undefined && resumeNoneResolver === undefined) {
+        this._resumeResolver = null;
+        this._resumeNoneResolver = null;
+        return;
+      }
+
+      if (resumeResolver && this._resumeResolver === resumeResolver) {
+        this._resumeResolver = null;
+      }
+      if (
+        resumeNoneResolver &&
+        this._resumeNoneResolver === resumeNoneResolver
+      ) {
+        this._resumeNoneResolver = null;
+      }
+    };
+
+    const finish = (
+      action: () => void,
+      resumeResolver?: ((data: { id: string }) => void) | null,
+      resumeNoneResolver?: (() => void) | null,
+      keepRequestId = false
+    ) => {
+      if (completed) return;
+      completed = true;
+      this._abortToolContinuation = null;
+      clearHandshakeResolvers(resumeResolver, resumeNoneResolver);
+      try {
+        action();
+      } catch {
+        // Stream may already be closed
+      }
+      if (requestId && !keepRequestId) {
+        activeIds?.delete(requestId);
+      }
+      streamController.abort();
+    };
+
+    const transport = this;
+
+    this._abortToolContinuation = () => {
+      if (completed) {
+        return false;
+      }
+
+      if (requestId === null) {
+        // Handshake hasn't completed yet — close the stream and clear
+        // resolvers so the subsequent onResume/handleStreamResuming
+        // becomes a no-op.
+        finish(
+          () => readerController.error(abortError),
+          onResumeRef,
+          onResumeNoneRef
+        );
+        return true;
+      }
+
+      try {
+        agent.send(
+          JSON.stringify({
+            type: MessageType.CF_AGENT_CHAT_REQUEST_CANCEL,
+            id: requestId
+          })
+        );
+      } catch {
+        // Ignore failures (e.g. agent already disconnected)
+      }
+
+      // keepRequestId=true: keep the ID in activeIds so onAgentMessage
+      // skips in-flight chunks until the server's done:true cleans it up
+      // (same pattern as sendMessages onAbort).
+      finish(
+        () => readerController.error(abortError),
+        onResumeRef,
+        onResumeNoneRef,
+        true
+      );
+      return true;
+    };
+
+    return new ReadableStream<UIMessageChunk>({
+      start(controller) {
+        readerController = controller;
+        let timeout: ReturnType<typeof setTimeout> | undefined;
+
+        const onResumeNone = () => {
+          if (timeout) clearTimeout(timeout);
+          finish(() => controller.close(), onResume, onResumeNone);
+        };
+
+        const onResume = (data: { id: string }) => {
+          if (requestId) return;
+
+          requestId = data.id;
+          activeIds?.add(requestId);
+          clearHandshakeResolvers(onResume, onResumeNone);
+          if (timeout) clearTimeout(timeout);
+
+          agent.send(
+            JSON.stringify({
+              type: MessageType.CF_AGENT_STREAM_RESUME_ACK,
+              id: requestId
+            })
+          );
+        };
+
+        onResumeRef = onResume;
+        onResumeNoneRef = onResumeNone;
+
+        timeout = setTimeout(
+          () => finish(() => controller.close(), onResume, onResumeNone),
+          5000
+        );
+
+        transport._resumeResolver = onResume;
+        transport._resumeNoneResolver = onResumeNone;
+        const onMessage = (event: MessageEvent) => {
+          try {
+            const data = JSON.parse(
+              event.data as string
+            ) as OutgoingMessage<UIMessage>;
+
+            if (
+              data.type !== MessageType.CF_AGENT_USE_CHAT_RESPONSE ||
+              requestId == null ||
+              data.id !== requestId
+            ) {
+              return;
+            }
+
+            if (data.error) {
+              finish(
+                () => controller.error(new Error(data.body || "Stream error")),
+                onResume,
+                onResumeNone
+              );
+              return;
+            }
+
+            if (data.body?.trim()) {
+              try {
+                const chunk = JSON.parse(data.body) as UIMessageChunk;
+                controller.enqueue(chunk);
+              } catch {
+                // Skip malformed chunk bodies
+              }
+            }
+
+            if (data.done) {
+              finish(() => controller.close(), onResume, onResumeNone);
+            }
+          } catch {
+            // Ignore non-JSON messages
+          }
+        };
+
+        const onClose = () => {
+          if (timeout) clearTimeout(timeout);
+          finish(() => controller.close(), onResume, onResumeNone);
+        };
+
+        agent.addEventListener("message", onMessage, {
+          signal: streamController.signal
+        });
+        agent.addEventListener("close", onClose, {
+          signal: streamController.signal
+        });
+
+        try {
+          agent.send(
+            JSON.stringify({
+              type: MessageType.CF_AGENT_STREAM_RESUME_REQUEST
+            })
+          );
+        } catch {
+          finish(() => controller.close());
+        }
+      },
+      cancel() {
+        if (requestId && transport.cancelOnClientAbort) {
+          transport.sendCancelFrame(requestId);
+          finish(() => {}, onResumeRef, onResumeNoneRef, true);
+        } else {
+          finish(() => {}, onResumeRef, onResumeNoneRef);
+        }
+      }
+    });
+  }
+
+  /**
+   * Creates a ReadableStream that receives resumed stream chunks
+   * and forwards them to useChat as UIMessageChunk objects.
+   */
+  private _createResumeStream(
+    requestId: string
+  ): ReadableStream<UIMessageChunk> {
+    // Read agent at resolve time (not when reconnectToStream was called)
+    // so chunk listener attaches to the latest socket after _pk changes.
+    const agent = this.agent;
+    const activeIds = this.activeRequestIds;
+    const chunkController = new AbortController();
+    const abortError = new Error("Aborted");
+    abortError.name = "AbortError";
+    let completed = false;
+
+    const finish = (
+      action: () => void,
+      keepId = false,
+      clearServerTurn = true
+    ) => {
+      if (completed) return;
+      completed = true;
+      if (clearServerTurn) {
+        this.clearActiveServerTurn(requestId);
+      }
+      try {
+        action();
+      } catch {
+        // Stream may already be closed
+      }
+      if (!keepId) {
+        activeIds?.delete(requestId);
+      }
+      chunkController.abort();
+    };
+
+    const cancelActiveRequest = () => {
+      if (completed) return false;
+      finish(() => streamController.error(abortError), true);
+      return true;
+    };
+    this.setActiveServerTurn(requestId, cancelActiveRequest);
+
+    let streamController!: ReadableStreamDefaultController<UIMessageChunk>;
+    const transport = this;
+
+    return new ReadableStream<UIMessageChunk>({
+      start(controller) {
+        streamController = controller;
+
+        const onMessage = (event: MessageEvent) => {
+          try {
+            const data = JSON.parse(
+              event.data as string
+            ) as OutgoingMessage<UIMessage>;
+
+            if (data.type !== MessageType.CF_AGENT_USE_CHAT_RESPONSE) return;
+            if (data.id !== requestId) return;
+
+            if (data.error) {
+              finish(() =>
+                controller.error(new Error(data.body || "Stream error"))
+              );
+              return;
+            }
+
+            // Parse and enqueue the chunk
+            if (data.body?.trim()) {
+              try {
+                const chunk = JSON.parse(data.body) as UIMessageChunk;
+                controller.enqueue(chunk);
+              } catch {
+                // Skip malformed chunk bodies
+              }
+            }
+
+            if (data.done) {
+              finish(() => controller.close());
+            }
+          } catch {
+            // Ignore non-JSON messages
+          }
+        };
+
+        const onClose = () => {
+          finish(() => controller.close(), false, false);
+        };
+
+        agent.addEventListener("message", onMessage, {
+          signal: chunkController.signal
+        });
+        agent.addEventListener("close", onClose, {
+          signal: chunkController.signal
+        });
+      },
+      cancel() {
+        if (transport.cancelOnClientAbort) {
+          transport.sendCancelFrame(requestId);
+          finish(() => {}, true);
+        } else {
+          finish(() => {}, false, false);
+        }
+      }
+    });
+  }
+}
