+++ packages/think/package.json
@@ -59,14 +59,22 @@
     "zod": "^4.4.3"
   },
   "peerDependencies": {
+    "@ai-sdk/react": "^3.0.204",
     "@chat-adapter/telegram": "^4.29.0",
     "agents": ">=0.16.0 <1.0.0",
     "ai": "^6.0.182",
+    "react": "^19.0.0",
     "vite": ">=6 <9",
     "zod": "^4.0.0"
   },
   "peerDependenciesMeta": {
+    "@ai-sdk/react": {
+      "optional": true
+    },
     "@chat-adapter/telegram": {
+      "optional": true
+    },
+    "react": {
       "optional": true
     },
     "vite": {
@@ -101,6 +109,10 @@
     "./messengers/telegram": {
       "types": "./dist/messengers/telegram.d.ts",
       "import": "./dist/messengers/telegram.js"
+    },
+    "./react": {
+      "types": "./dist/react.d.ts",
+      "import": "./dist/react.js"
     },
     "./tools/workspace": {
       "types": "./dist/tools/workspace.d.ts",
